linux-firmware: remove pointless LIC_FILES_CHKSUM entries
The point of the LIC_FILES_CHKSUM field is to alert the recipe maintainer
if the license statements in the sources have changed, so that the LICENSE
field can be updated.  A concrete example of this would be a COPYING that
says "This package is GPLv2" changing to "This package is GPLv3": the
LICENSE field would need to be updated from GPLv2 to GPLv3.

The linux-firmware package has tens of bespoke licenses contained in the
LICENSE.* files.  These files _are_ the license text and we use the file
name as the license name.  This means there's no purpose in having these
in LIC_FILES_CHKSUM as there's no action to be taken when the checksum
changes.

Signed-off-by: Ross Burton <ross.burton@arm.com>