dev-manual: update custom distribution section
In keeping with the addition of the motd message pointing out that
the poky DISTRO is a reference distribution, adjust the opening of the
Creating Your Own Distribution section to match. Additionally, add a
section on the end pointing out what users need to consider if they just
take a copy of the poky distribution and modify it.

Signed-off-by: Paul Eggleton <paul.eggleton@microsoft.com>