bitbake: cooker: Add an os.sync() call to revalidateCaches()
Unfortunately we've seen issues where changes are made by the UI side and are
in the disk cache but these tests don't pick up the file changes as they're
still proceeding through the VFS.

Any example is [YOCTO #15318] where logs showed a workspace and bbappend
being setup with devtool-source intercept but the server did not fully
pick up those changes leading to initial_rev (from devtool-source.bbclass)
being missing.

Add a sync call to try and avoid this despite the speed implications.

This was probably the kind of issue which led us to removing the inotify
code since that was also unsafe with concurrency since there was no way
to flush it.

(Bitbake rev: aafee36a8ceed37988d33fcc769dc85d66fbaa3e)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>