toolchain-scripts: export the target endianism and word size into meta-ide-support
Some build systems (notably, Meson) like to be told when cross-compiling
what the target endian and word size is.  We don't have these exported
via the variables, so add them using siteinfo as a source of data.

We already did this for normal SDKs, this does the same for meta-ide-support.

Signed-off-by: Ross Burton <ross.burton@arm.com>