meson: correct the host machine definition in SDKs
The SDK_ARCH is not the correct definiton for the host machine definition
inside the SDK, but because a nativesdk recipe doesn't know what the
final target will be these values should be set up at SDK installation
time via the environment script.

Put placeholders in the installed meson.cross file instead, and replace
them at SDK installation time with the correct values.

Signed-off-by: Ross Burton <ross.burton@arm.com>