pypi: rationalise variables
Improve the variables in the pypi recipe by being clearer about what
they are.

PYPI_PACKAGE is the name of the PyPI package: that is the name of the
package that is pip installed, or the title of the pypi page.

PYPI_PACKAGE for https://pypi.org/project/PyYAML/ is PyYAML.

https://peps.python.org/pep-0625/ says that a sdist should have a normalised
filename, so implement this normalisation to from PYPI_PACKAGE to get
the name of the sdist.  However, not all packages respect this PEP (in
particular it appears, old setuptools releases) so add PYPI_ARCHIVE_NAME
to be the name of the sdist.

Also update all the recipes that need amending.

Signed-off-by: Ross Burton <ross.burton@arm.com>