devtool/upgrade: rename RECIPE_UPDATE_EXTRA_TASKS -> RECIPE_UPGRADE_EXTRA_TASKS
'UPDATE' as a name is somewhat unfortunate as the variable is intended only for
the 'devtool upgrade' operation and devtool also has an 'update-recipe' operation.

I also intend to introduce additional variables in RECIPE_UPGRADE_* space,
so it's best to unify names now, while RECIPE_UPDATE_EXTRA_TASKS is very fresh.

Signed-off-by: Alexander Kanavin <alex@linutronix.de>