python3: Drop generating a static libpython
This currently puts a 60MB binary into the recipe sysroots in the native
case as well as the implications for the target. Since as far as I know,
we don't need/use this, stop generating it by adding a PACKAGECONFIG. It
can be enabled by anyone who does actually need to use it.

(From OE-Core rev: 477d5a5dadbb24e92908e8ec323bbb3a73ff3ed3)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>