prototype pydeps class
This class (which should be rolled into something else, like pep517) reads
the pyproject.toml file and parses the defined project dependencies.

It parses each requirement to be sure it's valid for the target and
generates a RDEPENDS using the virtual python-module-[package] names.
This should include the version field but there appears to be an issue
with opkg where it refused to resolve the version correctly so this is
currently disabled.

This needs either tomllib or tomli on the host, which is a blocker as we
don't assume that so I expect if we do this we'll need to vendor in tomli.

It also needs packaging.requirements (and possibly .markers) which I don't
believe is pulled in by any of our host dependencies, so we'll also need
to vendor or rewrite that.