fetch2: don't try to preserve everything when unpacking files
When copying files as part of the unpack we want to preserve the mode
(so executables are executable) and timestamps (for reproducibility) but
other fields are not relevant: as non-root the ownership won't be
preserved successfully anyway and there may be extended attributes that
are not able to be preserved (such as system.nfs4_acl if copying from
NFS to ext4).

Also expand the arguments to be clearer about what options are being
passed.

Signed-off-by: Ross Burton <ross.burton@arm.com>