bitbake: fetch2: don't try to preserve ownership when unpacking files
When copying files as part of the unpack we want to preserve the mode
(so executables are executable) and timestamps (for reproducibility) but
the ownership isn't really relevant: the files should be owned by the
build user and in situations where DL_DIR is shared between multiple
users there's no point asking to replicate ownership which cp will
silently fail to preserve.

Also expand the arguments to be clearer about what options are being
passed.

(Bitbake rev: ccea18d59ec4b66207733c7df5ee38880bb779bd)

Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Alexandre Belloni <alexandre.belloni@bootlin.com>