qa_image: move image_qa to between image and image_complete
Commit 62ce334e583ecdf1f93619f4131c0fa5d88d5b02 moved this after image as it was
noticed that do_image_qa is useless when rm_work is enabled as the rootfs
directory is deleted before image_qa is called.

However, this pactch made impossible the creation of a dependance from image_qa
to any image_* task where various types of images are generated. This disabled
the possibility to test a specific type of image (eg. a FIT image).

As rm_work is called as a postprocess of image_complete, qa_image can be done
after image as long as it is done before image_complete.

Signed-off-by: Louis Rannou <louis.rannou@non.se.com>