python3: Fix platform triplet detection
Currently, platform is based upon build_os detection which works ok
for native builds and where build_os and host_os are same i.e. glibc
based systems, but it fails for musl based systems to load correct
modules which are shared libraries e.g. python3-pydantic-core

This is already fixed upstream in 3.13+ but we need this backported
at minimum to get 3.12 or 3.11 working with musl

see
https://gitlab.alpinelinux.org/alpine/aports/-/issues/13227
https://github.com/meta-homeassistant/meta-homeassistant/issues/89
https://github.com/pypa/auditwheel/issues/349
https://github.com/PyO3/maturin/issues/1559

Signed-off-by: Khem Raj <raj.khem@gmail.com>