lib/oe/package: overwrite files from S with B when populating -src package
There's an interesting reproducibility problem triggered when a recipe
has a file in $S that is manipulated and then written into $B.  As a
concrete example, zlib has zconf.h in the source tree which is then
processed by configure to create zconf.h in the build tree.

When we populate the zlib-src package we have to work from the mapped
source paths in the binaries (of the form /usr/src/BP) and try to
identify where the file came from. To do this we iterate through the
mapping paths looking for files, and using cpio in a loop we end up with
the newest file with the name we're after in the -src package, which
tends to be the right thing to do.

However, when building zlib from scratch we create the -src package by
hard linking files from the source and build trees. Then as the final
act of do_package the sstate archive is created, and this clamps the
file timestamps to SOURCE_DATE_EPOCH (see fixtimestamp in sstate.bbclass).

The fundamental problem is that we clamp the mtimes in sstate_package in
a way that changes the source files. There are several paths through the
code:

1. Full build from clean
In this case B/zconf.h is freshly generated and has a current timestamp,
and the final zlib-src package has the correct file.

2. Repackage from existing tree
If we need to repackage but there is an existing build tree then we've
previously clamped the timestamp of zconf.h. When iterating the map paths
for files we first copy from $S and then $B, but as the file in $M has
an identical timestamp it isn't used. This results in the wrong file.

3. Repackage from sstate

[ YOCTO #15491 ]

Signed-off-by: Ross Burton <ross.burton@arm.com>