rust: Remove passing test cases from the exclude list in the Rust OE self-test.
We had previously excluded certain test cases due to failures in the MIPS64 architecture.
However, we are currently skipping MIPS64 for the Rust OE self-test. Therefore, we are
now including the test cases that pass for all other architectures except MIPS and MIPS64.

The rust-1.79 has been successfully tested with the current test inclusions.

Signed-off-by: Harish Sadineni <Harish.Sadineni@windriver.com>