bitbake: asyncrpc: timeout attempting to create a TCP connection
Timeout when trying to connect to an upstream that may be filtering
connections (instead of waiting endlessly) and raise a ConnectionError
which will allow the caller to note that the client is not connected.
This, for instance, ensures that a firewalled BB_HASHSERVE_UPSTREAM will
trigger warnings.

(Bitbake rev: 30d920644937dd16be8fb90c6aa59622fbdf6413)

Signed-off-by: Justin Bronder <jsbronder@cold-front.org>
Signed-off-by: Mathieu Dubois-Briand <mathieu.dubois-briand@bootlin.com>