classes/setuptools: do parallel builds in a clean directory
It turns out that by setting DIST_EXTRA_CONFIG you can tell setuptools to
read a supplementary configuration file. This lets us pass options that
we normally cannot reach via the bdist_wheel command.

Tell the build task to use a build directory outside of the source tree,
so that we can clean it before each rebuild.

Tell the build_ext task to do parallel builds, as it defaults to not. In
my testing this halves the build time of python3-pandas.

Signed-off-by: Ross Burton <ross.burton@arm.com>