classes-recipe/core-image: drop debug-tweaks IMAGE_FEATURE
Remove the 'debug-tweaks' IMAGE_FEATURE. It sounds friendly and kind to
developers, but it results in an image which root can login remotely
without a password.  This is incredibly useful for local development and
testing purposes, but we really want to be explicit that this is what is
happening instead of hiding it behind a vague "debug tweaks" statement.

To preserve the eixsting behaviour, debug-tweaks should be replaced with
these features:

  allow-empty-password empty-root-password allow-root-login

Signed-off-by: Ross Burton <ross.burton@arm.com>