systemd: split networkd into its own package
Fedora does this as well since systemd version 246.6-2:
https://src.fedoraproject.org/rpms/systemd/blob/rawhide/f/systemd.spec#_578
This proves that it is technically possible and officially supported by
systemd.

systemd is adding more and more default configuration files and examples.
For embedded devices where usually no adminstrator is using the examples
they should not be installed by default. Move them into a package which
is not installed by default.

Signed-off-by: Adrian Freihofer <adrian.freihofer@siemens.com>